email_spam_detector/requirements.txt
flask>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
shap>=0.42.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0

